Finished Prototype!
I got the prototype for the main mechanic of placing objects working. When holding an object, it raycasts for a placeable surface, and that surface then checks if it has a spot on its grid nearby that the object can fit in, and if it does, it lets you place it!